pyproject.toml
[project]
name = "booking-api"
version = "1.0.0"
description = "Sistema de reservas de salas de reunião"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8,<5.3",
    "djangorestframework>=3.16.1,<4.0",
    "djangorestframework-simplejwt>=5.5.1,<6.0",
    "drf-yasg>=1.21.11,<2.0",
    "psycopg2-binary>=2.9.11,<3.0",
    "dj-database-url>=3.0.1,<4.0",
    "django-cors-headers>=4.9.0,<5.0",
    "django-filter>=25.2,<26.0",
    "django-timezone-field>=7.1,<8.0",
    "python-decouple>=3.8,<4.0",
    "python-dateutil>=2.9.0,<3.0",
    "pytz>=2025.2",
    "tzdata>=2025.2",
    "gunicorn>=23.0.0,<24.0",
    "pyyaml>=6.0.3,<7.0",
    "markdown>=3.10,<4.0",
    "user-agents>=2.2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=6.1.0,<6.2",
    "ipython>=9.7.0,<9.8",
    "django-extensions>=4.1.0,<4.2",
]

test = [
    "pytest>=9.0.1,<9.1",
    "pytest-django>=4.11.1,<4.12",
    "pytest-cov>=7.0.0,<7.1",
    "factory-boy>=3.3.3,<4.0",
    "faker>=38.0.0,<39.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["settings", "booking", "core"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=.",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--tb=short",
    "--strict-markers",
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/wsgi.py",
    "*/asgi.py",
    "*/settings/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["*/migrations/*"]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]